Now the parsing of all queries except REGEX are done and tested.

+++ src/core_server/internal/ceql/value/integer_literal.hpp
@@ -6,11 +6,11 @@
 
 namespace InternalCORECEQL {
 struct IntegerLiteral final : public Value {
-  int value;
+  int64_t value;
 
   IntegerLiteral() noexcept {}
 
-  IntegerLiteral(int value) noexcept : value(value) {}
+  IntegerLiteral(int64_t value) noexcept : value(value) {}
 
   ~IntegerLiteral() noexcept {}
